tree-hard-graph-medium

+++ mycode/hot100/二叉树/437路径总和III.cpp
@@ -30,7 +30,7 @@ class Solution {
     // 前缀和
     unordered_map<long long, int> prefix;
     int res = 0;
-    void dfs(TreeNode* root, int sum, int cur_sum){
+    void dfs(TreeNode* root, int sum, long long cur_sum){
         if(root == nullptr){
             return;
         }
